+++ src/ru/nsu/ccfit/zuev/osu/helper/CentredSprite.java
@@ -7,8 +7,8 @@ public class CentredSprite extends Sprite {
 
     public CentredSprite(final float pX, final float pY,
                          final TextureRegion pTextureRegion) {
-        super(pX - pTextureRegion.getWidth() / 2, pY
-                - pTextureRegion.getHeight() / 2, pTextureRegion);
+        super(pX - (float) pTextureRegion.getWidth() / 2, pY
+                - (float) pTextureRegion.getHeight() / 2, pTextureRegion);
     }
 
 }
